+++ httphandler/downloadHandler.go
@@ -1,0 +1,71 @@
+package httphandler
+
+import (
+	"encoding/json"
+	"fmt"
+	"net/http"
+
+	"github.com/dgraph-io/badger/v3"
+	"github.com/gorilla/mux"
+)
+
+func (c Config) PlainDownloadHandler(w http.ResponseWriter, r *http.Request) {
+	vars := mux.Vars(r)
+	downloadKey := vars["downloadKey"]
+	param := vars["param"]
+
+	var jsonData []byte
+	err := c.Db.View(func(txn *badger.Txn) error {
+		item, err := txn.Get([]byte(downloadKey))
+		if err != nil {
+			return err
+		}
+		jsonData, err = item.ValueCopy(nil)
+		return err
+	})
+	if err != nil {
+		http.Error(w, "Data not found", http.StatusNotFound)
+		return
+	}
+
+	// Parse the JSON data to retrieve the specific parameter
+	paramMap := make(map[string]string)
+	if err := json.Unmarshal(jsonData, &paramMap); err != nil {
+		http.Error(w, "Error decoding JSON", http.StatusInternalServerError)
+		return
+	}
+
+	// Retrieve the value for the requested parameter
+	value, ok := paramMap[param]
+	if !ok {
+		http.Error(w, "Parameter not found", http.StatusNotFound)
+		return
+	}
+
+	// Return the value as plain text
+	w.Header().Set("Content-Type", "text/plain")
+	fmt.Fprintln(w, value)
+}
+
+func (c Config) DownloadHandler(w http.ResponseWriter, r *http.Request) {
+	vars := mux.Vars(r)
+	downloadKey := vars["downloadKey"]
+
+	var jsonData []byte
+	err := c.Db.View(func(txn *badger.Txn) error {
+		item, err := txn.Get([]byte(downloadKey))
+		if err != nil {
+			return err
+		}
+		jsonData, err = item.ValueCopy(nil)
+		return err
+	})
+	if err != nil {
+		http.Error(w, "Invalid download key or database error", http.StatusNotFound)
+		return
+	}
+
+	// Set header and write the JSON data to the response writer
+	w.Header().Set("Content-Type", "application/json")
+	w.Write(jsonData)
+}
