+++ middleware/config.go
@@ -1,0 +1,7 @@
+package middleware
+
+type Config struct {
+	MaxRequestSize     int64
+	RateLimitPerSecond float64
+	RateLimitBurst     int
+}
